Changed template calling convention
The previous way of specifying a template was to use a URI suffix
(e.g. .ajax, .html). This has the disadvantage of being
indistinguishable from the convention for specifying media type.

The new format uses a url directory convention e.g.:

/search/{template}/whatever

where template can be one of (full, light, etc).