Added an image on the metadata page displaying the extent of the data
The image provides some spatial context for the metadata, along with a
description of the bounding box coordinates. A link is also provided
to do a search for all other available data intersecting with that
extent.

The image currently links in to the OnEarth WMS Server
(http://onearth.jpl.nasa.gov/) to provide some background context, but
this can be replaced with other data sources.

There is also scope for showing a point on the earth instead of a
zoomed in view of the extent. This could be useful if the extent is
too small to provide much geographical context.