Added availability of metadata in KML format
Metadata can now be downloaded from the metadata page in KML
format. KML provides an excellent way of exploring the spatial context
of the metadata. The extents of the metadata are displayed along with
the abstract. A link is provided back to the HTML version where full
details can be accessed.