Enabled HTTP Etag caching for both results and metadata pages
An etag is created based on the metadata update date. For a result set
the update date is the most recent date in the results. This means
that there is less processing and network overhead if a client
re-requests a page that has not changed for them.

The etag also includes application version information so that when
the application is updated (e.g. the templates change) a client will
receive an updated resource even though the actual metadata content
may still be the same.