Enabled editing of search parameters
A link has been added to the catalogue results page that enables the
user to return to the search page and edit their current search. The
existing search parameters are propagated through the page so settings
are maintained.

Also, page information (i.e. page x of x) now appears in the titles
for catalogue results, providing useful navigational context.