Implement Etag HTTP caching for all template based pages
All pages generated by templates now automatically implement caching
based on Etags. This should speed up the application and make it more
cache friendly.

The etag is generated based on a combination of the template file path
and the application version number. Pages dealing with metadata
display that are also template based have the default etag
implementation overridden such that the metadata update date is used
instead of the template path.