Updated TileCache version to support X-Sendfile and expires HTTP headers
The TileCache application now returns the X-Sendfile HTTP header when
it receives a matching hit against a cached tile. This means that the
web server takes responsibility for delivering the tile to the
client. It does so more efficiently than tilecache, supplying
appropriate Etag and expires information. These settings are set in
the tilecache.cfg file. The expiry time is set to 3600 seconds (1
hour), meaning the browser will not request an updated copy of that
file within that time frame.

Apache must have X-Sendfile enabled with mod_xsendfile
(http://tn123.ath.cx/mod_xsendfile/). To enable it soley for the TMS,
the following Apache directive must be added:

  <Location /search/spatial/tms>
    XSendFile on
    XSendFileAllowAbove on
  </Location>