Filtered the TMS output through a GDAL VRT when creating the metadata image
The mapnik metadata image renderer uses the application TMS for its
background data. It accesses the TMS using the GDAL WMS
driver. Unfortunately the WMS driver does not specify colour
interpretations for the RGB raster bands in the received TMS data. As
there is no colour interpretation mapnik assumes the image is
grayscale; not what is required. Wrapping the TMS datasource in a GDAL
VRT allows us to add colour interpretations for the RGB bands which in
turn allows mapnik to interpret the image data properly.