Initial integration with SeaDataNet Vocabulary Web Service
The topic category (element 9) keywords are verified against the
vocabulary webservice. The returned data is used to include term and
definiton data for the keywords in the metadata details page.

Similarly data is extracted for the metadata keyword (element 11)
values. Unfortunately there doesn't seem to be a standard way to
determine the vocabulary list to which the keyword belongs, i.e. it is
difficult to link the keyword to the SeaDataNet vocabulary
list. Currently the code matches a pattern in the thesaurus name (Sub
Element 11.2.1) to determine the list to link to e.g. the name

SeaDataNet BODC Vocabulary (P011)

matches the vocabulary P011 by extracting the value from between the
brackets.

The portal connects up to the vocabulary web service using the SOAP
interface. See:

- http://www.seadatanet.org/standards_software/common_vocabularies
- http://www.bodc.ac.uk/products/web_services/vocab/