Added logging and error handling facilities to the application
The application can now emit messages detailing user errors
(e.g. incorrectly filled out fields). This is important in terms of
useability.

The user error handling interface is built on top of the standard
python logging module, and therefore provides a very flexible logging
interface to the application. This means that any application action
can be logged if required. Log messages can be sent to a variety of
different destinations if required, including email. In production,
for instance, if the application crashed unexpectedly it would be
useful to email the administrator with details of the crash.