Added dropdown select control for searching by country
The selecting a country from the dropdown in the geographic search
creates a bounding box around that country and zooms to the country
extent. The area can be cleared using the 'Clear area' button.

Performing a search after creating an area limits all results to the
spatial extent of the selected area.