Include summary of current search criteria
The search and results pages now include a box containing an
interpretation of the keyword search terms, the date range and the
geographic search (including the name of the search area).

This involved a number of changes to the application including placing
the areas in a sqlite database. This database is now the source of all
area information in the application (accessed via an internal Python
API as well as exposed via a REST API for AJAX related
functionality). This setup paves the way for adding more areas
e.g. ICES rectangles.