Added the ability to resize and drag an existing bounding box
The user can click on an existing bounding box on the spatial search
map which then highlights the box and disables map navigation using
the mouse. The user can now drag the box to a new location or resize
it using the handles in the corners. Altering the size of the box
updates the search criteria and result count.

Clicking on the box again de-selects the box and re-enables normal
navigation.