Made the geographic search more prominent
By default the geographic fieldset was collapsed. To make it more
prominent, but without incurring the loading penalty of showing the
entire map, this fieldset is by default expanded with an image and a
link that initialises the full geographic search if required.

The link and text makes use of tooltip style interface elements. This
style has also been copied to the results page to make the use of the
bookmark function more explicit.