Enabled user to edit search box coordinates
The user can now edit the coordinates of the search box by entering
decimal degrees in the format West,South,East,North into a text field
e.g.

  -8.623556,49.906189,1.759,60.84581

When this text field is edited the box on the map is updated to
reflect the new coordinates. Likewise, when the box on the map is
updated using other controls the text field is updated with the new
coordinates.