Added the ability to submit comments
A comment form now appears at the bottom of all HTML pages allowing
users to submit a comment along with their optional name and email
address. The page title, URL and user IP address are automatically
submitted as well.

The comment is emailed to the address provided in the new portal.ini
configuration file. This file also allows the SMTP server and port to
be specified.

This change also enhanced the logging such that log messages are
explictly split into errors, warnings and notices. This allows 'thank
you' messages to be passed through the message handling mechanism and
displayed appropriately.