Added HTML controls for searching by specific metadata fields
Clicking in the search term text box activates a panel with a
selection showing metadata fields to search. Choosing a field in turn
activates a dropdown of valid options or a free text search
box. Completing either of these controls populates the search term
field with the chosen search terms.