Fixed viewing of the light template in Internet Explorer
IE does not specifically specify a HTML content type that it accepts,
so for the light template receives the default application/xhtml+xml,
which it does not support. This fix adds middleware that checks to see
whether it is an IE browser requesting the resource, and updates the
environment to reflect that IE does specifically support text/html.