Enabled display of raw SOAP interactions between DWS and the portal
This facilitates debugging issues between the DWS and the Portal. This
is achieved by adding a HTTP GET parameter to a portal URL involving
the DWS. The GET parameter should be in the form of 'soap=request' or
'soap=response' the appropriate SOAP message is displayed in place of
the normal resource content.