Started integration with `python-skos`
`python-skos` is designed to replace the code that calls the NERC
vocabulary server on the fly.  Instead the vocabularies are cached and
accessed via the `python-skos` interface.  This increases speed and
robustness and enables all other vocabularies that are used but not on
the NERC server to be specified in SKOS format and added to the
vocabulary cache.