Patch SUDS to properly handle invalid service responses
The SUDS SOAP library has a bug whereby it fails to handle 500
internal server error responses that are not in SOAP format. This
fixes the problem locally. The issue has been flagged on the SUDS
issue tracker at <https://fedorahosted.org/suds/ticket/412>.

This replicates the fix committed in 4573474a8c9d836.